adds FenixDetails placeholder
